==> 1.read.py <==
# 可读性指标
import jieba
import re
import pandas as pd
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

df = pd.read_excel('data_mda.xlsx') 

chinese_char_map = {}
with open('./chinese_unicode_table.txt', 'r', encoding='UTF-8') as f:
    lines = f.readlines()
    for line in lines[6:]:
        line_info = line.strip().split()
        chinese_char_map[line_info[0]] = int(line_info[6])  # 将笔画数转换为整数

def chinese_text_statistics(text):
    # 分词，并过滤掉数字
    words = [word for word in jieba.cut(text) if not re.match(r'\d+', word)]

    # 以句号、问号、感叹号作为分隔符，切分句子
    sentences = re.split('[。？！]', text)

    # 去除空句子
    sentences = [s for s in sentences if s.strip()]

    word_count = len(words)  # 统计字数
    sentence_count = len(sentences)  # 统计句数
    avg_sentence_length = word_count / sentence_count if sentence_count > 0 else 0  # 计算平均句长
    
    # 遍历文本中的每个汉字，查找其对应的笔画数
    stroke_counts = [chinese_char_map.get(char, 0) for char in text if char in chinese_char_map]

    # 计算平均笔画数
    average_stroke_count = sum(stroke_counts) / len(stroke_counts) if len(stroke_counts) > 0 else 0

    # 计算被动句数和转折连词数
    passive_sentence_count = 0
    concessive_conjunction_count = 0
    for sentence in sentences:
        # 判断是否包含被动语态的迹象（以“被”开头）
        if re.search(r'(?<![a-zA-Z0-9_])受|被|让|遭受', sentence):
            passive_sentence_count += 1

        # 判断是否包含常见的转折连词
        if re.search(r'(?<![a-zA-Z0-9_])但|然而|但是|虽然|尽管|却|只是|不过|至于|不料|岂知|偏偏|只是|不过如此|尽管如此|不过|如此', sentence):
            concessive_conjunction_count += 1

    # 使用外部语料库进行匹配：新华字典和MBA智库
    uncommon_words_corpus = set(['匕', '刁', '丐', '歹','戈','夭','仑','讥','冗','邓','艾','夯','凸','卢','叭','叽','皿','凹','囚','矢','乍','尔','冯','玄','邦','迂','邢','芋','芍','吏','夷','吁','吕','吆','屹','廷','迄','臼','仲','伦','伊','肋','旭','匈','凫','妆','亥','汛','讳','讶','讹','讼','诀','弛','阱','驮','驯','纫','玖','玛','韧','抠','扼','汞','扳','抡','坎','坞','抑','拟','抒','芙','芜','苇','芥','芯','芭','杖','杉','巫','杈','甫','匣','轩','卤','肖','吱','吠','呕','呐','吟','呛','吻','吭','邑','囤','吮','岖','牡','佑','佃','伺','囱','肛','肘','甸','狈','鸠','彤','灸','刨','庇','吝','庐','闰','兑','灼','沐','沛','汰','沥','沦','汹','沧','沪','忱','诅','诈','罕','屁','坠','妓','姊','妒','纬','玫','卦','坷','坯','拓','坪','坤','拄','拧','拂','拙','拇','拗','茉','昔','苛','苫','苟','苞','茁','苔','枉','枢','枚','枫','杭','郁','矾','奈','奄','殴','歧','卓','昙','哎','咕','呵','咙','呻','咒','咆','咖','帕','账','贬','贮','氛','秉','岳','侠','侥','侣','侈','卑','刽','刹','肴','觅','忿','瓮','肮','肪','狞','庞','疟','疙','疚','卒','氓','炬','沽','沮','泣','泞','泌','沼','怔','怯','宠','宛','衩','祈','诡','帚','屉','弧','弥','陋','陌','函','姆','虱','叁','绅','驹','绊','绎','契','贰','玷','玲','珊','拭','拷','拱','挟','垢','垛','拯','荆','茸','茬','荚','茵','茴','荞','荠','荤','荧','荔','栈','柑','栅','柠','枷','勃','柬','砂','泵','砚','鸥','轴','韭','虐','昧','盹','咧','昵','昭','盅','勋','哆','咪','哟','幽','钙','钝','钠','钦','钧','钮','毡','氢','秕','俏','俄','俐','侯','徊','衍','胚','胧','胎','狰','饵','峦','奕','咨','飒','闺','闽','籽','娄','烁','炫','洼','柒','涎','洛','恃','恍','恬','恤','宦','诫','诬','祠','诲','屏','屎','逊','陨','姚','娜','蚤','骇','耘','耙','秦','匿','埂','捂','捍','袁','捌','挫','挚','捣','捅','埃','耿','聂','荸','莽','莱','莉','莹','莺','梆','栖','桦','栓','桅','桩','贾','酌','砸','砰','砾','殉','逞','哮','唠','哺','剔','蚌','蚜','畔','蚣','蚪','蚓','哩','圃','鸯','唁','哼','唆','峭','唧','峻','赂','赃','钾','铆','氨','秫','笆','俺','赁','倔','殷','耸','舀','豺','豹','颁','胯','胰','脐','脓','逛','卿','鸵','鸳','馁','凌','凄','衷','郭','斋','疹','紊','瓷','羔','烙','浦','涡','涣','涤','涧','涕','涩','悍','悯','窍','诺','诽','袒','谆','祟','恕','娩','骏','琐','麸','琉','琅','措','捺','捶','赦','埠','捻','掐','掂','掖','掷','掸','掺','勘','聊','娶','菱','菲','萎','菩','萤','乾','萧','萨','菇','彬','梗','梧','梭','曹','酝','酗','厢','硅','硕','奢','盔','匾','颅','彪','眶','晤','曼','晦','冕','啡','畦','趾','啃','蛆','蚯','蛉','蛀','唬','啰','唾','啤','啥','啸','崎','逻','崔','崩','婴','赊','铐','铛','铝','铡','铣','铭','矫','秸','秽','笙','笤','偎','傀','躯','兜','衅','徘','徙','舶','舷','舵','敛','翎','脯','逸','凰','猖','祭','烹','庶','庵','痊','阎','阐','眷','焊','焕','鸿','涯','淑','淌','淮','淆','渊','淫','淳','淤','淀','涮','涵','惦','悴','惋','寂','窒','谍','谐','裆','袱','祷','谒','谓','谚','尉','堕','隅','婉','颇','绰','绷','综','绽','缀','巢','琳','琢','琼','揍','堰','揩','揽','揖','彭','揣','搀','搓','壹','搔','葫','募','蒋','蒂','韩','棱','椰','焚','椎','棺','榔','椭','粟','棘','酣','酥','硝','硫','颊','雳','翘','凿','棠','晰','鼎','喳','遏','晾','畴','跋','跛','蛔','蜒','蛤','鹃','喻','啼','喧','嵌','赋','赎','赐','锉','锌','甥','掰','氮','氯','黍','筏','牍','粤','逾','腌','腋','腕','猩','猬','惫','敦','痘','痢','痪','竣','翔','奠','遂','焙','滞','湘','渤','渺','溃','溅','湃','愕','惶','寓','窖','窘','雇','谤','犀','隘','媒','媚','婿','缅','缆','缔','缕','骚','瑟','鹉','瑰','搪','聘','斟','靴','靶','蓖','蒿','蒲','蓉','楔','椿','楷','榄','楞','楣','酪','碘','硼','碉','辐','辑','频','睹','睦','瞄','嗜','嗦','暇','畸','跷','跺','蜈','蜗','蜕','蛹','嗅','嗡','嗤','署','蜀','幌','锚','锥','锨','锭','锰','稚','颓','筷','魁','衙','腻','腮','腺','鹏','肄','猿','颖','煞','雏','馍','馏','禀','痹','廓','痴','靖','誊','漓','溢','溯','溶','滓','溺','寞','窥','窟','寝','褂','裸','谬','媳','嫉','缚','缤','剿','赘','熬','赫','蔫','摹','蔓','蔗','蔼','熙','蔚','兢','榛','榕','酵','碟','碴','碱','碳','辕','辖','雌','墅','嘁','踊','蝉','嘀','幔','镀','舔','熏','箍','箕','箫','舆','僧','孵','瘩','瘟','彰','粹','漱','漩','漾','慷','寡','寥','谭','褐','褪','隧','嫡','缨','撵','撩','撮','撬','擒','墩','撰','鞍','蕊','蕴','樊','樟','橄','敷','豌','醇','磕','磅','碾','憋','嘶','嘲','嘹','蝠','蝎','蝌','蝗','蝙','嘿','幢','镊','镐','稽','篓','膘','鲤','鲫','褒','瘪','瘤','瘫','凛','澎','潭','潦','澳','潘','澈','澜','澄','憔','懊','憎','翩','褥','谴','鹤','憨','履','嬉','豫','缭','撼','擂','擅','蕾','薛','薇','擎','翰','噩','橱','橙','瓢','磺','霍','霎','辙','冀','踱','蹂','蟆','螃','螟','噪','鹦','黔','穆','篡','篷','篙','篱','儒','膳','鲸','瘾','瘸','糙','燎','濒','憾','懈','窿','缰','壕','藐','檬','檐','檩','檀','礁','磷','瞭','瞬','瞳','瞪','曙','蹋','蟋','蟀','嚎','赡','镣','魏','簇','儡','徽','爵','朦','臊','鳄','糜','癌','懦','豁','臀','藕','藤','瞻','嚣','鳍','癞','瀑','襟','璧','戳','攒','孽','蘑','藻','鳖','蹭','蹬','簸','簿','蟹','靡','癣','羹','鬓','攘','蠕','巍','鳞','糯','譬','霹','躏','髓','蘸','镶','瓤','攥','矗'])
    accounting_terms_corpus = set(['对账','对应账户','定期清查','定期盘存制','订本式账簿','调整账户','调整分录','单式记账凭证','单式记账法','从属账户','成本计算账户','财产清查','簿记','不定期清查','补充登记法','表外账户','备抵账户','备抵附加账户','备查账簿','序时账簿','一次凭证','银行存款日记账','永续盘存制','原始凭证','暂记账户','增减记账法','债权结算账户','债权债务结算账户','债务结算账户','账户','账户编号','账户对应关系','账项调整','专用记账凭证','转回分录','资金来源账户','资产负债账户','转账凭证','资金运用账户','自制原始凭证','总分类账簿','总分类账户','附加账户','付款凭证','分类账簿','多栏式日记账核算形式','结账','结账分录','借贷记账法','局部清查','卡片式账簿','跨期摊提账户','累计凭证','联合账簿','明细分类账簿','明细分类账户','盘存账存','平行登记','全面清查','日记总账','日记总账核算形式','三式记账法','实账户','试算表','试算平衡','收付记账法','收款凭证','损益表账户','通用记账凭证','通用日记账核算形式','外来原始凭证','现金日记账','虚账户','汇总原始凭证','汇总记账凭证核算形式','工作底稿','复式记账凭证','复式记账法','复合分录','划线更正法','汇总原始凭证','会计凭证','会计科目表','会计科目','红字更正法','会计核算形式','过账','会计分录','会计循环','会计账簿','活页式账簿','集合分配账户','计价对比账户','记账方法','记账规则','记账凭证','记账凭证核算形式','记账凭证汇总表核算形式','简单分录','结算账户','期间费用','收入的确认','公司债券发行价格','固定资产折旧','可转换债券','公司债券利息摊销','加速折旧法','营业外收支净额','公司债券利率','应收账款出借','无担保债券','后进先出法',',,其他货币资金','应付票据贴现','先进先出去','在发建工程','固定资产更换与固定资产改良','实地盘存制','收益总括观点','损益表法','可变现净值法','应付福利费','基本业务利润','固定资产扩建','债券赎回','应交折基金应收账款出售','或有负债','销货退回与折让','零售价格法','现金折扣','特定履行法','其他业务利润','公司债券','销售法','应付票据','认股权','固定资产修理','有担保债券','销售费用','应付股利','基本生产应收票据','无形资产','收款法','所得税','流动负债','生产法','计划成本核算','废弃和生置法','盘存法','流动资产','购货折扣','商誉','应收账款','投资收益','营业利润','预提费用','股本','企业管理费','公司债券偿还','坏账','固定资产重估价','银行存款','固定资产','利润总额','利润分配','应计费用','商标权','全部履行法','净利润','应付利润','未分配利润','收益债券','货币资金','利息资本化','法定公益金','工程物资','预付账款','其他应收款','现金','预收账款','公司债券发行','应付工资','实收资本','盈余公积','管理费用','／土地使用权','股利','应交税金','流动资金','负商誉','费用的确认','短期投资','短期借款','递延资产','低值易耗品','/当期经营观点','待摊费用','待核销基建支出待处理流动资产损失','待处理固定资产损失','存货销售的影响','折旧折旧方法','折旧率','支出','直线法','职工福利基金专项拨款专利权','住房基金','重置成本法','专项物资专项资产专有技术','专营权','资本公积','资产负债表法','资金占用和资金来源','自然资源','存货','车间经费偿债基金','长期应付款','长期投资','长期借款','长期负债','财务费用','拨定留存收益','标准成本法','变动成本法','比例履行法','包装物','版权','独立董事','市场附加值',',投资中心','利润中心','酌量性费用中心','收入中心','合并前利润','现金分配计划','安全付款表','合并每股收益','期货交易市场','期货交易','举债经营融资租赁','金融工具','企业集团','年度报告','内部往来','合伙企业','合并资产负债表','合并主体的所得税会计美','合并现金流量表','合并价差','合并会计报表','购买法','企业整体价值','权益结合法','期内所得税分摊美','期末存货的未实现损益','公司间的长期资产业务','名义货币保全','基金论','功能性货币美','汇兑损益','合并财务状况变动表','换算损益','举债经营收购美','母公司持股比例变动','交互分配法美','货币项','合伙清算','控股合并','关税','名义货币单位','记账本位币','经营租赁','流动非流动性法','/经济利润','破产受托人清算组会计','联合会计报表','权益法','共同费用分配','货币非货币法','/利率期货交易','简单权益法','汇率','母公司','红利法','库藏股法美','劳务因素','精算报告','全面分摊法','固定资产投资方向调节税','合并费用','间接标价法','买入汇率','期货合约','混合合并','控投公司','股票指数期货','横向销售','固定汇率','纳税影响法','记账汇率','横向合并','合并前股利','可变现净值','企业合并会计','平仓盈亏','卖出汇率','金融期货交易','会计利润','合并损益表','公允价值','期权','间接控股','两笔交易观','破产清算','企业合并','企业论','商品寄销','权益理论','融资租赁','商品期货交易','商誉法','生产能力保全','升水','少数股东损益','少数股东权益','上市公告书','剩余权益论','时态法','实体理论','实体论','受托人美','特定变动','所得税会计','所得税的跨期分摊','税务会计','售后回租','耕地占用税','个人所得税','个人财务报表美','改组计划美','改组','复杂权益法','附属公司','负权人偿金','浮动汇率','分支机构会计','推定赎回损益','推定赎回','投机','贴水','特定物价指数','分支机构','分期收款销货','分次清算','分部报告','房地产收入','房地产成本','房地产','多种汇率法','对境外实体的净投资','订量单位','递延法','当代理论','单一汇率法','退休金','退休金会计美','退休金给付义务美','退休金成本净额美','退休基金资产美','流转会计','合伙权益的转让美','购买力损益','非货币性项目','单行合并美','外汇','外币会计报表','外币折算风险','外币统账法','外币套期保值','外币投资风险','外币资产风险','外币会计报表折算','外币兑换风险','外币承诺','外币负债风险','外币持有风险','外币分账法','外币','外币业务','吸收合并','物价变动会计','无偿债能力','完全合并','物价指数','物价变动','完全应计法','物价总指数','外汇期货交易','下推会计美','先折算后调整法','现行成本/稳值货币会计','现行成本','现行成本会计','先调整后折算法','销售代理处','相互持股','相对账户调节','新合伙人入伙','向上销售','衍生金融工具','销售式融资租赁','向下销售','消费税','一笔交易观','业主权论','一般物价水准会计','一般购买力单位','一般购买力保全','印花税','应付税款法','营业亏损抵免','以外币表示的应收款项或应付款项','一次总付清算','营业税','永久性差异','原合伙人退伙','原始成本','远期汇率','时间性差异','暂时性差异','增值表','债权人会议','债务重整','账面汇率','直接融资租赁','直接标价法','直接控股','招股说明书','中间汇率','中期报告','重置成本','转租赁','准改组美','资本保全','资本化价值','资本因素','资产负债法','/存货转让价格','创立合并','出租人会计','持有/产损益','持仓盈亏','承租人会计','成本回收法','纵向合并','综合变动','子公司权益变动','子公司','资源税','成本法','财产信托会计美','财产税','部分分摊法','不合并子公司','最低退休金负债美','租赁','租金','表融资','比例合并美','保证金法','资本资产定价模型','成本会计关联业务往来','国有资本经营预算制度','能源审计','投资中心','利润中心','酌量性费用中心','收入中心','固定成本','标准成本中心','作业成本计算法','直接人工成本差异','直接材料成本差异','在产品计价','联产品成本计算','生产成本汇总程序','制造费用差异','实际成本与估计成本','工资费用分配','成本曲线','农业生产成本','原始成本和重置成本','工程施工成本','直接成本与间接成本','可控成本','制造费用分配','理论成本与应用成本','辅助生产成本分配','期间费用','成本控制程序','成本记录',',成本计算分批法','成本计算分步法','直接人工成本差异','成本控制方法','内河运输成本','生产费用要素','历史成本与未来成本','可避免成本与不可避免成本','成本计算期','平均成本与个别成本','社会成本','铁路运输成本','废品损失','单位成本与总成本','成本开支范围','沿海运输成本','成本转账','全面成本控制','商品销售成本','价格差异','汽车运输成本','存置成本','已耗成本与未耗成本','航空运输成本','相关成本与非相关成本','因素分析法','目标成本','成本计算分类法','远洋运输成本','定额成本','跨期摊提费用分配','计划成本','数量差异','燃料费用分配','定额成本控制制度','定额管理','可递延成本与不可递延成本','成本控制标准','副产品成本计算','责任成本','生产损失核算','生产成本','预计成本','成本结构','房地产开发成本','主要成本与加工成本','决策成本','成本计算品种法','在产品成本','工厂成本','成本考核','制造费用','动力费用分配','趋势分析法','成本计算简单法','责任成本层次','对比分析法','约当产量比例法','原始记录','可比产品成本分析','预算成本','销售成本','停工损失','等级产品成本计算','分散核算','宏观经济成本','综合费用分配','全部成本','集中核算','商品采购成本','成本考核指标','闲置成本','账面成本','主要产品单位成本分析','再生产成本','增量成本','成本控制','全部产品成本分析','成本流程','内部成本报表','成本计算方法','成本计算对象','成本计算单位','成本计划完成情况分析','成本计划管理体系','成本计划','成本会计','成本核算原则','成本核算程序','成本核算成本','成本核算','成本归集','成本管理','成本分析','成本分配','成本分类账','成本分类','成本费用界限','成本调整','成本差异','成本报告','成本','车间成本','厂内经济核算制','厂内结算价格','产品寿命周期成本','产品成本项目','产品成本技术经济分析','产品成本计划','产品成本','产成品成本','财务成本与管理成本','材料费用分配','不可控成本','标准成本控制制度','标准成本','比率分析法','报告成本','半成品成本','企业财务会计管理层收购','预算实践应注意的问题','平衡计分卡','经济附加值','热钱','财务总监','预算外资金','企业财务','期权市场','期货市场','可转让定期存单市场','货币市场','黄金市场','国有独资公司','股份有限公司','股份两合公司','公司','二级市场','独资企业','店头市场','承兑市场','拆借市场','财务制度','财务政策','财务预测','财务控制','金融市场','财务决策','财务监督','财务计划','财务活动','财务管理组织','一级市场','无限责任公司','外汇市场','贴现市场','企业组织形式','财务管理职能','财务管理原则','财务管理内容','财务管理目标','财务管理环境','证券交易所','证券市场','资本市场','有限责任公司','财务管理环节','财务管理对象','财务管理','财务分析','配比原则','旅游、饮食服务企业会计','施工企业会计','民航运输企业会计','企业会计','商品流通企业会计','权责发生制原则','农业会计','实现原则','历史成本原则','外商投资企业会计','通用报表','铁路运输企业会计','所有者权益','所有者权益','实质重于形式','修正性惯例','信息系统论','相关性原则','微观会计','客观性原则','可比性原则','谨慎性原则','金融企业会计','交通运输企业会计','建设单位会计','记账本位币','计量属性','及时性原则','货币计量','会计准则','会计主体','会计职业道德','会计职能','会计预测','会计要素','会计研究','会计学科体系','会计学','会计信息','会计任务','会计人员','会计确认','会计目标','会计理论结构','会计理论','会计控制','会计决策','会计监督','会计假设','会计记录','会计计量','会计机构','会计环境','会计核算','会计管理体制','会计分期','会计对象','会计等式','会计本质','会计报表','宏观会计','会计','汇总报表','划分资本性支出与收益性支出原则','合并报表','管理活动论','管理会计','管理工具论','股份制企业会计','公认会计原则','公共会计','工业会计','个别报表','高新技术企业会计','负债','费用','反馈价值','对外经济合作企业会计','对外报表','对内报表','一致性原则','艺术论','房地产开发企业会计','邮电通信企业会计','预测价值','真实与公允','持续经营','成本报表','财务会计原则','财务会计概念框架','财务会计','政府及非营利组织会计','重要性原则','专用报表','资产','资金','资金运动','财务报告','财务报表要素','财务报表','币值稳定假设','保险企业会计','政府及非盈利组织会计预算周转金','预算举借债务','预算结余','中国预算会计','预拨款项','医院资产','医院支出','医院预算管理办法','医院收入','医院结余','医院基金','医院会计制度','医院会计科目','医院会计报表分析','医院会计报表','医院会计','医院负债','医院财务制度','一般预算支出','一般预算收入','行政单位资产','行政单位支出','行政单位暂付款','行政单位暂存款','行政单位预算管理办法','行政单位收入','行政单位年终清理和结转','行政单位净资产','行政单位经费支出','行政单位结余','行政单位会计制度','行政单位会计报表分析','行政单位会计','行政单位固定基金','行政单位负债','行政单位财务规则','行政单位拨入经费','事业收入','事业单位资产','事业单位专用基金','事业单位支出','事业单位预算管理方式','事业单位应缴款项','事业单位收入','事业单位流动资产','事业单位净资产','事业单位借入款项','事业单位结余','事业单位基金','事业单位会计准则','事业单位会计','事业单位固定资产','事业单位固定基金','事业单位负债','事业单位对外投资','事业单位财务清算','上缴上级支出','上级补助收入','其他收入','科学事业单位资产','科学事业单位支出','科学事业单位预算','科学事业单位收入','科学事业单位结余','科学事业单位会计制度','科学事业单位会计报表分析','科学事业单位会计','科学事业单位成本费用管理','科学事业单位财务制度','经营支出','经营收入','基金预算支出','基金预算收入','基金预算结余','国家预算','国家决算','高等学校资产','高等学校支出','高等学校预算管理方式','高等学校收入','高等学校净资产','高等学校结余','高等学校会计制度','高等学校会计报表分析','高等学校会计','高等学校负债','高等学校财务制度','附属单位缴款','对附属单位补助','财政总预算会计制度','财政总预算会计年终清理','财政总预算会计年终结账','财政总预算会计','财政资产','财政周转金支出','财政周转金收入','财政周转基金','财政支出','财政性存款','资金调拨收入','专用基金支出','专用基金收入','专用基金结余','中华人民共和国预算法','资金调拨支出','财政收入','财政净资产','财政负债','财政补助收入','拨入专款'])

    # 计算次常用字数
    uncommon_words_count = sum(1 for char in text if char in uncommon_words_corpus)

    # 计算会计术语的次数
    accounting_terms_count = sum(1 for word in words if word in accounting_terms_corpus)

    return word_count, sentence_count, avg_sentence_length, average_stroke_count, passive_sentence_count, concessive_conjunction_count, uncommon_words_count, accounting_terms_count

# # 示例文本
# text = "这是一个示例文本。其中包含一些被动语态的句子，也包含一些转折连词。真是歹毒的对账峻峻峻。例如但是、尽管等。文本还包含一些次常用字和会计术语，例如会计术语1、次常用字2等。"

# # 调用函数并输出结果
# word_count, sentence_count, avg_sentence_length, average_stroke_count, passive_sentence_count, concessive_conjunction_count, uncommon_words_count, accounting_terms_count = chinese_text_statistics(text)
# print("字数:", word_count)
# print("句数:", sentence_count)
# print("平均句长:", avg_sentence_length)
# print("平均笔画数:",average_stroke_count)
# print("被动句数:", passive_sentence_count)
# print("转折连词数:", concessive_conjunction_count)
# print("次常用字数:", uncommon_words_count)
# print("会计术语次数:", accounting_terms_count)

# 对每个文本计算八个指标，并添加新列
df[['word_count', 'sentence_count', 'avg_sentence_length', 'average_stroke_count',
    'passive_sentence_count', 'concessive_conjunction_count',
    'uncommon_words_count', 'accounting_terms_count']] = df['text'].apply(lambda x: pd.Series(chinese_text_statistics(x)))

df.to_excel('output_data_mda.xlsx', index=False)  # 根据需要修改输出文件的路径和格式

# df = pd.read_excel('output_data_mda.xlsx') 

# 选择八个指标的列
selected_columns = ['word_count', 'sentence_count', 'avg_sentence_length', 'average_stroke_count',
                     'passive_sentence_count', 'concessive_conjunction_count',
                     'uncommon_words_count', 'accounting_terms_count']

# 标准化数据
scaler = StandardScaler()
df_scaled = scaler.fit_transform(df[selected_columns])

# 执行主成分分析
pca = PCA()
df_pca = pca.fit_transform(df_scaled)

# 创建新的 DataFrame 包含主成分
df_pca_result = pd.DataFrame(data=df_pca, columns=[f'PC{i+1}' for i in range(df_pca.shape[1])])

# 将主成分 DataFrame 与原始 DataFrame 合并
df_with_pca = pd.concat([df, df_pca_result], axis=1)

# 获取每个主成分的贡献率
explained_variance_ratio = pca.explained_variance_ratio_

# 计算每个主成分的权重
weights = explained_variance_ratio / explained_variance_ratio.sum()

# 计算每个文本的主成分加权平均：前六个，已经将近90%
df_with_pca['Readty'] = (df_with_pca['PC1'] * weights[0] +
                         df_with_pca['PC2'] * weights[1] +
                         df_with_pca['PC3'] * weights[2] +
                         df_with_pca['PC4'] * weights[3] +
                         df_with_pca['PC5'] * weights[4] +
                         df_with_pca['PC6'] * weights[5]
                        )

df_with_pca.to_excel('output_data_mda_with_readty.xlsx', index=False)  # 根据需要修改输出文件的路径和格式
